Move DummyStore into the sixlowpan dummy code
